src/personal_assistant_agency/requirements.txt
agency-swarm>=0.4.4
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
pytz>=2024.1
python-dotenv>=1.0.0
beautifulsoup4 